+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bug fixes
 
 - fix command palette highlights wrong item on the list
+- react to cursor shape changes in shell, attach and plugin views
 
 ## 0.5.1 - 2026-07-01